deployment/backend/requirements.txt
pandas
flask
tensorflow-cpu
numpy
gunicorn